requirements.txt
# Core AI and Computer Vision
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0

# Audio and GUI
pygame>=2.0.0
gtts>=2.2.0
tkinter  # Usually comes with Python

# Data Processing
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
pyyaml>=5.4.0
requests>=2.25.0
tqdm>=4.62.0

# Optional: GPU Support
# torch-cuda  # Uncomment if using CUDA

# Development and Testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
